requirements.txt
requests
pandas
streamlit
matplotlib
plotly
python-dotenv
google-generativeai 